Add missing shortcut tooltips to control buttons
Those shortcuts were already working before, but not really
discoverable for users.

To achieve a consistent UI and keep the wording in line with existing
conventions and future shortcuts (i.e. no whole sentences, shortcut in
parenthesis), the mute tooltip is changed again to occupy two lines (to
limit the width of the tooltip), to be less wordy (so it is faster to
read), to use the "mute" term (to help memorizing the M shortcut, at
least in English), and to align with the parenthesis style.