Fix broken Return key in export route dialog
2a431932 made it possible to press `Return` to confirm the export dialog.
After 5b52abc1 this stopped working though, even when the text input had focus.

This is due to moving the `<input>` out of the `<form>` and referencing the
form's `name` attribute instead. However "The value of [the form] attribute
must be the id of a <form> in the same document" according to
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button.

Adding such an `id` restores the `Return` key to working properly.

The same change is done for the no-go form, where the same issue would otherwise
affect a later patch adding similar `Return` key handling.

Test Plan:
- Create Route, open "Export route" dialog.
- Move focus to an input field.
- Press `Return`.
- The file dialog for saving should open.