Normalize `surface` and `smoothness` variants
The `surface` tag exists in different variants,
e.g. `surface`, `cycleway:surface` etc.

Previously, the `surface` and `smoothness` tags were only processed for route analysis if they were found in their canonical form in the BRouter server response.

With this commit, the variants are normalized down to the main tag name which has the effect that they're included in the route analysis.

Fixes #438