Fix URL for calling OSM
The OSM URL in the bottom right-hand corner contains "karte.html" in the path. If the link is called, OSM throws away the ".html" part and redirects to the start page without parameters.

In order for the parameters to be properly adopted by OSM, the URL would have to be "https://www.openstreetmap.de/karte?zoom=12&lat=53.5138&lon=7.8207&layers=B000TF", for example.